[lldb][NFC] Moved the SharedSocket class to Socket.*
This is prerequisite for #104238.
